Add LevelObjectFactory and LevelParser for level object management
- Implement LevelObjectFactory to create Character objects from level definitions, handling resource paths and scaling based on group adjustments.
- Introduce LevelParser to parse level JSON data, extracting object definitions, group adjustments, and other relevant properties.
- Enhance JSON extraction utility functions for better handling of strings, integers, and floats.
- Ensure proper handling of percentage-based positioning and scaling for level objects.
